requirements.txt
Django
django-storages[s3]
python-decouple
Pillow